Change the metrics interface
Previously it was about getting data and letting the circuit breaker make the decision. Instead, let's allow the metrics to make the decision. This makes it so the metrics can control things a little more which means resilient is more customizeable. For example, let's say someone wanted success to always close the circuit and failure to always open it. For this scenario, someone can now just create a new metrics interface that responds to success, failure, reset, and the two under methods where instead of faking error percentage to be 0 (closed) or 100 (open), they can now just return true/false for under_error_threshold_percentage? based on if there are any errors or not and make success reset the metrics.